tweaks

+++ MMOCoreORB/src/server/zone/objects/tangible/weapon/WeaponObjectImplementation.cpp
@@ -349,7 +349,7 @@ void WeaponObjectImplementation::fillAttributeList(AttributeListMessage* alm, Cr
 
 	// Force Cost
 	if (getForceCost() > 0)
-		alm->insertAttribute("forcecost", (int)getForceCost());
+		alm->insertAttribute("forcecost", (float)getForceCost());
 
 	for (int i = 0; i < getNumberOfDots(); i++) {
